added directive for word and character limit validation
Its little tricky to place a word or character based validation to UI-Grid. But with a little understanding of ui-grid events, this can be achieved. I have used BEGIN_CELL and EVENT_START and EVENT_STOP events from the ui-grid library to suitably place a validation on cell. This is tested with IE 11 and Chrome with ui-grid - v3.2.6